pyproject.toml
[project]
name = "vlm-exp"
version = "0.1.0"
description = "VLM experiments with Ollama for image captioning and metadata extraction"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pyexiftool>=0.5.6",
    "moondream-station>=0.1.0",
    "moondream>=0.1.0",
    "pillow>=10.0.0",
    "gliner>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
